Soil table handling
Added stand alone module for parsing and writing soil tables in summa. It seems to work okay. 

to use the code, you can use put the module in the same working directory as your python script, or use the sys.path.append() method to add a custom path to your python code